A lot of files have been added, but don't worry. A lot of them were added with the scaffold command. These are automagically created by a plugin, were changed by me because they didn't work, and are now the backbone of the song industry. They just need to have more content.
The template works as expected, but now I just want it to be fancy.

On the main page I expect the user to be able to add a new song, and it will look, by ajax, if the song is already added. After the user submits a song he will go to the page of the song, where he can submit a youtube video link, which will be embedded in the song. Every time someones goes to this page and no youtube link has been added a youtube search will be done where the artist's name and the title will be submitted, after which the viewer is presented by a couple of youtube videos and will select the correct video. After the correct video has been added it will be stored in the database so that it won't be a hassle to look it up.

This is the main submitting phase. This currently has the highest priority.

Then the voting phase. There will be a special user, VoteForSimon or something, who, once logged in, will see a special overview, where all the songs which needs voting are presented. The presenter has an option to embed a song, which will show the youtube/soundcloud link, or a search result or something like that, and will select. Then the voter can either 'upvote' or 'downvote' the song, which will result in the song to be purchased later, or not at all. This will all be done in a visually appealing way.

That was the second highes thing that needed to be done.

Lastly, I want to make it easy for the person who buys the songs in order to keep track of the state of a song. It will only show the songs that were voted, but not yet bought. This is quite easy to do, but I will have to do it nonetheless.

Anywho, this is a long commit, and if you don't like it, deal with it!