Spotify resuls are looking a lot better now. The same logic of youtube is applied here, I even bought the spotify icon. :D. Anywho, selecting and/or deselecting stuff doesn't work, but it's close.
I actually want to create some sort of interface model for the searchproviders. This for the voting page, so they know how to display the song, depending on the provider. (Spotify results are displayed differently). Anywho, I'll be rewriting the spotify methodoligy soon, using PHP as a pre-processor for the spotify json results. :)