Added more functionality. Login and register pages are ready, and I must say that they are gorgeous, which isn't that hard with bootstrap. :P. Anywho, made the layouts even MORE extensible, because there are more files/snippets which can be used. These provice the basic structure, i.e. the header, navbar and javascripts. This was needed because I didn't want to restructure the whole thing for the signin/signup page.
I want to users to be able to signup so that they can receive updates about their songs. This might not be what is wanted, but I wanted it nonetheless. This also makes it easier for us to vote and/or add songs to our top 5.

So, this was quite an update. I will try to make logging in / registering a thing for this evening. :D